chore: release to GitHub always fails (#264)
GitHub always complains that the files we try to attach to
the release already exist. Add the `--clobber` flag to overwrite,
which will stop the error.

(I'm not quite sure why the files are already there, and it seems
to be inconsistent as well. But this will at least fix the error.)