Use AL2023 Lambda runtime
As the Go runtime is deprecated.
As part of this change, the compiled Go binary main is renamed
to bootstrap, as that is the name that the AL2023 Lambda runtime
expects.
We now build using the standard Go Docker image, as the AWS Lambda
Go image is deprecated.
It's important to note that the Dockerfile in the lambda/ folder
is not used as part of the runtime at all. It's only function is
to produce the /asset/bootstrap binary that is then uploaded to
Lambda and used with the AL2023 runtime. This was the case before
this change, except it used to produce /asset/main to use with
the GO runtime.