fix: Use Amazon Linux 2 provided runtime (#424)
* Use AL2023 Lambda runtime

As the Go runtime is deprecated.
As part of this change, the compiled Go binary main is renamed
to bootstrap, as that is the name that the AL2023 Lambda runtime
expects.
We now build using the standard Go Docker image, as the AWS Lambda
Go image is deprecated.
It's important to note that the Dockerfile in the lambda/ folder
is not used as part of the runtime at all. It's only function is
to produce the /asset/bootstrap binary that is then uploaded to
Lambda and used with the AL2023 runtime. This was the case before
this change, except it used to produce /asset/main to use with
the GO runtime.

* Rename _GOPROXY to GOPROXY

Now we're no longer building using the AWS build image, there's
no conflict with that image if we use GOPROXY.

* Fix Release lambda Github step

The rename from main to bootstrap was missed from this step.

* Use golang:1 as the build Go image

So we're tracking the latest 1.x Go version, like we were when
we were using public.ecr.aws/sam/build-go1.x:latest.

* Update major version

Update the major version to 3, as changing the base build image
is a breaking change. Run `npx projen`.

* Update readme

* update tasks.json

---------

Co-authored-by: Parker Scanlon <69879391+scanlonp@users.noreply.github.com>