feat: Add support to explicitly specify a CPU architecture for the to-be-copied image instead of relying on auto-detection (#961)
### What's the problem?

It's observed that when copying a multi-arch source image from location
A to B in ECR, we always end up with `amd64` in the destination image at
location B. The hypothesis is that the underlying lambda function used
to auto-detect the architecture is backed up by Dockerfile with [a
hard-coded
runtime.GOARCH](https://github.com/cdklabs/cdk-ecr-deployment/blob/dcd74cdfc394d6a43a5517cb5cce1843e1fe3111/lambda/Dockerfile#L11).

### Proposed Solution
Instead of solely relying on an auto-detection mechanism, this PR is
meant to add a feature to allow users to explicitly specify a CPU
architecture for the to-be-copied image, by triggering the underlying
logic
[here](https://github.com/containers/image/blob/5263462aa97ae23d3a620b4dcdbbc557665a42c8/internal/pkg/platform/platform_matcher.go#L164).
This will unblock the usage of Graviton instances for [Amazon
SageMaker](https://aws.amazon.com/sagemaker/) and other computing
platforms.

As a side note, I've run `npm run release` on my dev box to refresh
`API.md`

---------

Co-authored-by: Bo Xiong <xiongbo@amazon.com>
Co-authored-by: Rico Huijbers <rix0rrr@gmail.com>