fix: failing prebuilt download creates a 0 byte bootstrap file (#1097)
When the download of the prebuilt lambda bootstrap file fails for
whatever reason, an empty file is created in it's place. While the
initial execution will abort, this empty file stays in place and
subsequent executions assume it's a valid file just by nature of being
there.

We change this approach to be more resilient and delete the file in case
of any failure.