feat: prebuilt lambda code is now bundled with the package (#1098)
This construct sometimes comes with a prebuilt Lambda, and sometimes
downloads the lambda code from GitHub whenever it synths.

The reason is historical:

- In the past, the Lambda always used to be downloaded from GitHub at
synth time
- We changed that to download from GitHub at packaging time, with an
opportunistic fallback to downloading at synth time if the binary wasn't
there

In this PR, we're simplifying the situation a bunch, and removing the
ability to download at synth time. Instead, the Lambda gets built at
build time, and included into the package. This is an unconditional part
of the build, and it must succeed for the build to succeed. We therefore
never have to download again, because we know the Lambda will always be
in the package.

Also removes the in-source Lambda build; we're now splitting directories
into `lambda-src` and `lambda-bin`.

This construct still has the ability to fall back to a local synth.
We're going for that next 😉

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>