feat!: simplify construct by removing build-on-synth (#1103)
This removes the ability to build the Lambda source at synthesis time.
Instead, the Lambda is built at package time, and the prebuilt package
is always used at runtime.

This also removes the ability to configure the Lambda runtime, entry
point, and environment variables, because those configuration options
aren't necessary anymore if the Lambda image is fixed.

Since we are removing options, this is a breaking change.

While we are bumping the major version number anyway, the name of the
custom resource was `Custom::CDKBucketDeployment`, which this PR changes
to `Custom::CDKECRDeployment` for accuracy's sake.

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>